requirements.txt
discord.py==2.6.4
langgraph==1.0.5
pydantic==2.12.5
PyYAML==6.0.2
python-dotenv==1.0.1
aiohttp==3.10.5
beautifulsoup4==4.12.3
playwright==1.57.0
